chore: move shared objects to shared.py
