+++ sqlmesh/core/engine_adapter/databricks.py
@@ -6,7 +6,7 @@ import typing as t
 import pandas as pd
 from sqlglot import exp
 
-from sqlmesh.core.engine_adapter.base import CatalogSupport, InsertOverwriteStrategy
+from sqlmesh.core.engine_adapter.shared import CatalogSupport, InsertOverwriteStrategy
 from sqlmesh.core.engine_adapter.spark import SparkEngineAdapter
 from sqlmesh.core.schema_diff import SchemaDiffer
 from sqlmesh.utils import classproperty
